Move pictures into public folder to enable cache mechanism.
