+++ components/MembershipTabs/Achievements.tsx
@@ -1,9 +1,9 @@
 import Image from "next/image";
 import styles from "./Achievements.module.css";
-import AchieveFacePic1 from "../../assets/Membership/Achievements/AchieveFakePic1.svg";
-import AchieveFacePic2 from "../../assets/Membership/Achievements/AchieveFakePic2.svg";
-import AchieveFacePic3 from "../../assets/Membership/Achievements/AchieveFakePic3.svg";
-import AchieveFacePic4 from "../../assets/Membership/Achievements/AchieveFakePic4.svg";
+import AchieveFacePic1 from "../../public/assets/Membership/Achievements/AchieveFakePic1.svg";
+import AchieveFacePic2 from "../../public/assets/Membership/Achievements/AchieveFakePic2.svg";
+import AchieveFacePic3 from "../../public/assets/Membership/Achievements/AchieveFakePic3.svg";
+import AchieveFacePic4 from "../../public/assets/Membership/Achievements/AchieveFakePic4.svg";
 
 export const Achievements = () => {
   return (
