+++ pages/index.tsx
@@ -1,17 +1,17 @@
 import styles from "../styles/Home.module.css";
-import AdsImage from "../assets/index.ads.svg";
-import FakePoint from "../assets/index.fakepoint.svg";
-import IndexMidTab1 from "../assets/IndexMidTab/midtab.myticket.svg";
-import IndexMidTab2 from "../assets/IndexMidTab/midtab.arrivedtime.svg";
-import IndexMidTab3 from "../assets/IndexMidTab/midtab.roadinternet.svg";
-import IndexMidTab4 from "../assets/IndexMidTab/midtab.coupon.svg";
-import IndexMidTab5 from "../assets/IndexMidTab/midtab.achievement.svg";
-import OnlineMallPic1 from "../assets/IndexOnlineMallPic/onlinemall.pic1.svg";
-import OnlineMallPic2 from "../assets/IndexOnlineMallPic/onlinemall.pic2.svg";
-import OnlineMallPic3 from "../assets/IndexOnlineMallPic/onlinemall.pic3.svg";
-import SpacesPic1 from "../assets/IndexSpacesPic/spaces.pic1.svg";
-import SpacesPic2 from "../assets/IndexSpacesPic/spaces.pic2.svg";
-import SpacesPic3 from "../assets/IndexSpacesPic/spaces.pic3.svg";
+import AdsImage from "../public/assets/index.ads.svg";
+import FakePoint from "../public/assets/index.fakepoint.svg";
+import IndexMidTab1 from "../public/assets/IndexMidTab/midtab.myticket.svg";
+import IndexMidTab2 from "../public/assets/IndexMidTab/midtab.arrivedtime.svg";
+import IndexMidTab3 from "../public/assets/IndexMidTab/midtab.roadinternet.svg";
+import IndexMidTab4 from "../public/assets/IndexMidTab/midtab.coupon.svg";
+import IndexMidTab5 from "../public/assets/IndexMidTab/midtab.achievement.svg";
+import OnlineMallPic1 from "../public/assets/IndexOnlineMallPic/onlinemall.pic1.svg";
+import OnlineMallPic2 from "../public/assets/IndexOnlineMallPic/onlinemall.pic2.svg";
+import OnlineMallPic3 from "../public/assets/IndexOnlineMallPic/onlinemall.pic3.svg";
+import SpacesPic1 from "../public/assets/IndexSpacesPic/spaces.pic1.svg";
+import SpacesPic2 from "../public/assets/IndexSpacesPic/spaces.pic2.svg";
+import SpacesPic3 from "../public/assets/IndexSpacesPic/spaces.pic3.svg";
 import Image from "next/image";
 
 const Ads = () => {
